deepgram-cli/requirements.txt
python-dotenv
deepgram-sdk
deepgram-captions
